Add layout, splitting, MDC, and thread name support to Logback appender
Custom layouts may be specified using Logback "Layout" objects.

Long messages are split into pieces with a configurable prefix on the
second through Nth pieces, to avoid truncation due to the log message
length limit. If possible, the splitting is done at newline boundaries
so as not to chop lines in half in stack traces or other multi-line
messages.

MDC properties and thread names are automatically included as
attributes on log messages.